pyproject.toml
[project]
name = "agnoagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno==1.8.0",
    "chainlit>=2.9.5",
    "ddgs>=9.10.0",
    "duckduckgo-search>=8.1.1",
    "google-genai>=1.58.0",
    "groq>=1.0.0",
    "lancedb>=0.26.1",
    "mcp>=1.25.0",
    "mcp-server-calculator>=0.2.0",
    "openai>=2.15.0",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0.45",
    "wikipedia-mcp>=1.7.0",
    "yfinance>=1.0",
]
